% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_series.R
\name{compare_series}
\alias{compare_series}
\title{Compare series from a CSO table}
\usage{
compare_series(...)
}
\arguments{
\item{...}{series-objects.}
}
\value{
An mts object.
}
\description{
Compare series from a CSO table
}
\examples{
#library(statbanker)
#library(plotly)
#NAQ01 <- getStatBankData("NAQ01", type = "px")
#PCR_SA <- get_series(NAQ01,2)
#GCR_SA <- get_series(NAQ01,4)
#ITR_SA <- get_series(NAQ01,6)
#compare_series(PCR_SA@data[,"Y-on-Y"],ITR_SA@data[,"Y-on-Y"],GCR_SA@data[,"Y-on-Y"])
}
