% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_differences.R
\name{compute_differences}
\alias{compute_differences}
\title{Calculate first differences of a series}
\usage{
compute_differences(series_ts, lag)
}
\arguments{
\item{series_ts}{A ts object.}

\item{lag}{An integer value.}
}
\value{
A ts object.
}
\description{
Calculate first differences of a series
}
\examples{
library(statbanker)
NAQ01 <- getStatBankData("NAQ01", type = "px")
PCR_SA <- extract_series(NAQ01, index = 2)
compute_differences(PCR_SA, 1)
}
