% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_series.R
\name{extract_series}
\alias{extract_series}
\title{Extract series from a CSO table}
\usage{
extract_series(table_px, index)
}
\arguments{
\item{table_px}{A px object from statbanker package.}

\item{index}{An integer or string to locate the series.}
}
\value{
A ts (time-series) object.
}
\description{
Extract series from a CSO table
}
\examples{
library(statbanker)
NAQ01 <- getStatBankData("NAQ01", type = "px")
PCR_SA <- extract_series(NAQ01, index = 2)
}
