% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_series.R
\name{find_series}
\alias{find_series}
\title{Find series in a CSO table}
\usage{
find_series(text, table_px, ...)
}
\arguments{
\item{text}{A string of text (i.e. the search term).}

\item{table_px}{A px object from statbanker package.}

\item{...}{Additional arugments}
}
\value{
A list containing the matched series and index numbers.
}
\description{
Find series in a CSO table
}
\examples{
library(statbanker)
NAQ01 <- getStatBankData("NAQ01", type = "px")
find_series("exports", NAQ01)
}
