% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_series.R
\name{list_series}
\alias{list_series}
\title{List series in a CSO table}
\usage{
list_series(table_px)
}
\arguments{
\item{table_px}{A px object from statbanker package.}
}
\value{
A character vector.
}
\description{
List series in a CSO table
}
\examples{
library(statbanker)
NAQ01 <- getStatBankData("NAQ01", type = "px")
list_series(NAQ01)
}
